Add root element title
Typically we've hidden most of the "root element" display stuff. Change that to emphasize the key or title of the root document, give it a special underline, and just generally make it behave like other rows in terms of selecting (will make it easy to copy/paste the root element, too).

Also:
* Give documentation stories titles to display
* Update tests to handle root element being less of a special case in terms of selection
* Code styling here and there.