Add schema prop to type definitions
I'm trying out a new approach to schemas: associating them with definitions. The idea being custom definitions will tend to have a specific schema in mind, and so this makes it explicit and relatively simple to provide them to treema and use them in the passed-in schema.

Updated the example custom definitions to this new approach, and some other minor fixes.